# api/openapi.yaml
openapi: 3.0.3
info:
  title: Photobox API
  description: Complete API specification for Photobox backend - a photo management system
  version: 1.0.0
  contact:
    name: Photobox API Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.photobox.example.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
      description: PASETO token authentication. Use the /api/login endpoint to obtain a token.
  schemas:
...
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Photo unique identifier
        name:
          type: string
          description: Photo name
        filesystemPath:
          type: string
          description: Path on filesystem
        sourcePath:
          type: string
          description: Original source path
        albumId:
          type: string
          format: uuid
          description: Parent album ID
        tags:
          type: string
          description: Photo tags
        metadata:
          type: object
          description: Photo metadata (EXIF, etc.)
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        createdEpoch:
          type: integer
          description: Creation epoch time
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        thumbnail:
          type: string
          format: byte
          description: Thumbnail binary data (optional)
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        data:
          type: object
          description: Response data (varies by endpoint)
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        data:
          type: array
          items: {}
        metadata:
          type: object
          properties:
            fromId:
              type: string
              description: Base64 encoded starting epoch
            toId:
              type: string
              description: Base64 encoded ending epoch
            count:
              type: integer
              description: Number of items returned
            nextPage:
              type: string
              description: URL to next page
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        messages:
          type: array
          items:
            type: string
          description: Array of error messages
paths:
...
  /api/photo/thumbnail/{id}:
    get:
      tags:
        - Photos
      summary: Get photo thumbnail
      description: Retrieve photo thumbnail as binary image data
      operationId: getPhotoThumbnail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Photo UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thumbnail retrieved successfully
          content:
            image/*:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
        '404':
          description: Photo not found
...
  /api/photos:
    get:
      tags:
        - Photos
      summary: List photos
      description: Get paginated list of photos for a specific album
      operationId: listPhotos
      security:
        - bearerAuth: []
      parameters:
        - name: albumId
          in: query
          required: true
          description: Album UUID to filter photos
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 10
        - name: fromId
          in: query
          description: Base64 encoded epoch for pagination
          schema:
            type: string
        - name: thumbnail
          in: query
          description: Include thumbnail data in response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Photos retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Photo'
        '400':
          description: Missing required albumId parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/photos/index:
    post:
      tags:
        - Photos
      summary: Start photo indexing
      description: Start asynchronous photo indexing job
      operationId: indexPhotos
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Indexing job started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Indexing job already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...